test update again

+++ app/src/main/java/hexlet/code/dto/urls/UrlsPage.java
@@ -14,7 +14,7 @@ import java.util.Map;
 
 public class UrlsPage extends BasePage {
     private List<Url> urls;
-    private long pageNumber;
+    private int pageNumber;
     private List<UrlCheck> lastCheck;
     private String conditionNext;
     private String conditionBack;
